Numberplate-predict
A number plate is a unique identifier assigned to a vehicle, typically consisting of letters and numbers. It is issued by a government authority for registration and legal purposes. The format varies by country, with some using a combination of numbers, letters, and symbols. It serves as a way to track, identify, and regulate vehicles on the road.